requirements.txt
blueprint==3.4.2
Flask==2.0.2
Flask-Mail==0.9.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2